+++ api/__init__.py
@@ -1,0 +1,4 @@
+from api.app import app
+
+if __name__ == '__main__':
+    app.run(debug=True)
