+++ api/author/db.py
@@ -1,0 +1,12 @@
+all_authors = [
+    {
+        'id': 1,
+        'nickname': 'TolstoyProd',
+        'name': 'Лев Толстой'
+    },
+    {
+        'id': 2,
+        'nickname': 'PushkinProd',
+        'name': 'Александр Пушкин'
+    }
+]
