+++ api/book/db.py
@@ -1,0 +1,26 @@
+all_books = [
+    {
+        'id': 1,
+        'title': 'Война и мир',
+        'annotation': 'Про книгу война и мир',
+        'author_id': 1
+    },
+    {
+        'id': 2,
+        'title': 'Евгений Онегин',
+        'annotation': 'Про книгу Евгений Онегин',
+        'author_id': 2
+    },
+    {
+        'id': 3,
+        'title': 'Анна Каренина',
+        'annotation': 'Про книгу Анна Каренина',
+        'author_id': 1
+    },
+    {
+        'id': 4,
+        'title': 'Полтава',
+        'annotation': 'Про книгу Полтава',
+        'author_id': 2
+    }
+]
