+++ install.sh
@@ -4,3 +4,4 @@ python3 -m venv venv
 . venv/bin/activate
 pip install Flask
 pip install Flask-HTTPAuth
+export FLASK_APP=api
